ai_utils.py
import google.generativeai as genai
import os

# Configure the API key
genai.configure(api_key=os.getenv('GEMINI_API_KEY'))

# Set up the model
model = genai.GenerativeModel('gemini-pro')

def get_gemini_response(prompt):
    response = model.generate_content(prompt)
    return response.text

def generate_embedding(text):
    # This is a placeholder. You'll need to use a model that can generate embeddings.
    # For example, you might use a different model or API for this purpose.
    # The exact implementation will depend on the embedding model you choose.
    pass