requirements.txt
streamlit==1.27.0
google-generativeai==0.3.1
pandas==2.1.1
plotly==5.17.0
python-dotenv==1.0.0
pinecone-client==2.2.2